+++ AMBroSIA/src/game/BonusDrop.java
@@ -19,7 +19,7 @@ public class BonusDrop extends MapObject {
     
     BonusDrop(int[] coordinates, GameState gameState, int type)
     {
-        super(new int[] {0, 0} , 0, coordinates, 0, gameState);
+        super(new float[] {0, 0} , 0, coordinates, 0, gameState);
         this.type = type;
         this.ttl = TIME_TO_LIVE;
     }
